% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfastExt.R
\name{bfastExt}
\alias{bfastExt}
\title{bfastExt}
\usage{
bfastExt(x, start, monend = NULL, cells = NULL, sensor = NULL,
  interactive = FALSE, plot = FALSE, order = 1, h = 0.25,
  level = 0.05, mc.cores = 1, magnThresh = 0, magnPeriod = 1,
  verbose = FALSE, ...)
}
\arguments{
\item{x}{Rasterbrick with time-series data}

\item{start}{(see bfastmonitor in bfast package)}

\item{monend}{(see bfastmonitor in bfast package)}

\item{cells}{Numeric vector of cells on which to run the function}

\item{sensor}{(see bfastmonitor in bfast package)}

\item{interactive}{(see bfastmonitor in bfast package)}

\item{plot}{(see bfastmonitor in bfast package)}

\item{order}{(see bfastmonitor in bfast package)}

\item{h}{(see bfastmonitor in bfast package)}

\item{level}{(see bfastmonitor in bfast package)}

\item{mc.cores}{Number of cores to be used.}

\item{magnThresh}{NDMI threshold}

\item{magnPeriod}{Period (in years)}

\item{verbose}{(see bfastmonitor in bfast package)}
}
\value{
List of information
}
\description{
Multicore version of extended BFAST function. Should be replaced with B. DeVries's bfmThresh in due time.
}
\author{
E. van Schaik
}

