% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfastOpt.R
\name{bfastOpt}
\alias{bfastOpt}
\title{bfastOpt}
\usage{
bfastOpt(data = NULL, training.set = NULL, nIter = NULL)
}
\arguments{
\item{A}{param}
}
\value{
... Test
}
\description{
Some test description
}
\author{
E. van Schaik
}

